Refactor and clean up PalFinder service and spec
This commit refactors the PalFinder service and its corresponding spec to be more efficient and adheres to best practices.
The code now uses chain scopes and the 'pluck' method for performance improvement and readability.
It also restructures the spec to use better function descriptions and replace hardcoded values with factories and FFaker for better reusability and randomness in the test suite.

Included is a minor change in the database schema to reflect a more appropriate attribute name, 'interest', in the Interest model.

Finally, the FFaker gem is added for generating fake data in tests, which helps randomize the test cases and create a larger data set.